Add support for SSL engine loading the private key.
Added option 'key-engine' specifying the name of the engine that
will load the private key.

For example this can be "tpm" to use the OpenSSL TPM engine module
(libengine-tpm-openssl in Debian).

It defaults to the built-in UI methods because openssl-tpm-engine
doesn't yet support user data being sent to the callback functions.
A patch for that on its way to them.

Signed-off-by: Thomas Habets <habets@google.com>